F01: Add implementation of CategorieService
