+++ src/main/java/com/carto/sn/service/CategorieServiceImpl.java
@@ -1,5 +1,42 @@
 package com.carto.sn.service;
 
-public class CategorieServiceImpl {
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+
+import com.carto.sn.dao.CategorieRepository;
+import com.carto.sn.entities.Categorie;
+
+public class CategorieServiceImpl implements CategorieService{
+	
+	@Autowired
+	private CategorieRepository categorieRepository;
+
+	@Override
+	public List<Categorie> toutesLesCategories() {
+		return categorieRepository.findAll();
+	}
+
+	@Override
+	public Categorie ajoutCategorie(String nomCategorie) {
+		Categorie categorie = null;
+		categorie = categorieRepository.save(new Categorie(nomCategorie));
+		return categorie;
+	}
+
+	@Override
+	public Categorie findByNomCategorie(String nomCategorie) {
+		return categorieRepository.findByNomCategorie(nomCategorie);
+	}
+
+	@Override
+	public Categorie findByIdCategorie(Long idCategorie) {
+		return categorieRepository.findById(idCategorie).orElse(null);
+	}
+
+	@Override
+	public void supprimerCategorie(Long idCategorie) {
+		categorieRepository.deleteById(idCategorie);		
+	}
 
 }
